// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/memory.ts
export enum MemoryRecordKind {
    /**
     * @generated from protobuf enum value: MEMORY_RECORD_KIND_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: MEMORY_RECORD_KIND_EPISODIC = 1;
     */
    EPISODIC = 1,
    /**
     * @generated from protobuf enum value: MEMORY_RECORD_KIND_SEMANTIC = 2;
     */
    SEMANTIC = 2,
    /**
     * @generated from protobuf enum value: MEMORY_RECORD_KIND_OBSERVATIONAL = 3;
     */
    OBSERVATIONAL = 3
}
/**
 * @generated from protobuf enum nimi.runtime.v1.MemoryCanonicalClass
 */
export enum MemoryCanonicalClass {
    /**
     * @generated from protobuf enum value: MEMORY_CANONICAL_CLASS_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: MEMORY_CANONICAL_CLASS_NONE = 1;
     */
    NONE = 1,
    /**
     * @generated from protobuf enum value: MEMORY_CANONICAL_CLASS_PUBLIC_SHARED = 2;
     */
    PUBLIC_SHARED = 2,
    /**
     * @generated from protobuf enum value: MEMORY_CANONICAL_CLASS_WORLD_SHARED = 3;
     */
    WORLD_SHARED = 3,
    /**
     * @generated from protobuf enum value: MEMORY_CANONICAL_CLASS_DYADIC = 4;
     */
    DYADIC = 4
}
// ...
export enum MemoryReplicationOutcome {
    /**
     * @generated from protobuf enum value: MEMORY_REPLICATION_OUTCOME_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: MEMORY_REPLICATION_OUTCOME_PENDING = 1;
     */
    PENDING = 1,
    /**
     * @generated from protobuf enum value: MEMORY_REPLICATION_OUTCOME_SYNCED = 2;
     */
    SYNCED = 2,
    /**
     * @generated from protobuf enum value: MEMORY_REPLICATION_OUTCOME_CONFLICT = 3;
     */
    CONFLICT = 3,
    /**
     * @generated from protobuf enum value: MEMORY_REPLICATION_OUTCOME_INVALIDATED = 4;
     */
    INVALIDATED = 4
}